Don't use static Semaphores
The data types used to identify the various POSIX constructs are
normally the size of a pointer and do not contain the entire data
required to implement the functionality.

Some software implicitly depends on the small size of these data
structures when storing them, e.g. pjsip storing mutexes into an
internal pool in the pools implementation.

This patch changes the usage of static semaphores to heap allocated ones
in an attempt to make the data structures used smaller.